Move tests to relevant files (#1771)
* move tests to diff files


* codacy errors


* unused imports


* minor csm removal
